feat: 完成环境构建优化
