+++ demo/package.json
@@ -10,13 +10,13 @@
   },
   "dependencies": {
     "ahooks": "^3.7.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "showdown": "^2.1.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.2",
     "@types/showdown": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "@vitejs/plugin-react-swc": "^3.0.0",
