+++ src/content/index.ts
@@ -1,6 +1,6 @@
 import logger from '../utils/log'
-import { extractArticle, extractTokenDOMs } from './preprocess'
 import { TOC } from './toc'
+import { loadTOC } from './load'
 
 let toc: TOC
 let cachedLoadPattern = '0'
@@ -11,19 +11,9 @@ let cachedLoadPattern = '0'
   logger.info('content script load 前 获得到的用户配置', { loadPattern })
   cachedLoadPattern = loadPattern
   if (loadPattern === '0') return
-  load()
+  toc = loadTOC()
 })()
 
-function load() {
-  const article = extractArticle()
-  const tokenDOMs = extractTokenDOMs(article)
-
-  logger.info({ article, tokenDOMs })
-  if (!tokenDOMs?.length) {
-    logger.info('there is no article / headings')
-  }
-  toc = TOC.createTOC(tokenDOMs, {})
-}
 
 let lastUrl = location.href
 setInterval(() => {
@@ -40,14 +30,14 @@ function onUrlChange() {
     if (cachedLoadPattern === '0') {
       return
     }
-    load()
+    loadTOC()
   }, 200)
 }
 
 chrome.runtime.onMessage.addListener((command) => {
   logger.info('content script received command', command)
   if (!toc) {
-    load()
+    loadTOC()
   } else {
     toc.toggle()
   }
