+++ src/utils/log.ts
@@ -3,7 +3,7 @@ import { isDev } from './env'
 class Logger {
   info(...args) {
     if (!isDev) return
-    console.log(...args)
+    console.log("[awesome-toc]",...args)
   }
 }
